Add intentional support for `SafeHaskell`
This is the first version whose `SafeHaskell` properties have become
an intentional part of the API contract; previous versions were merely
accidentally safe-inferred (or not depending on various factors; in
other words, this was a fragile property). If you rely on
`SafeHaskell` to consider module imports from `parsec` *safe*, this is
the first version of `parsec` which actually guarantees a well-defined
state.

Support for GHC 7.0, GHC 7.2, and GHC 7.4.1 has been dropped,
consequently the support window starts with GHC 7.4.2.

This addresses #88